+++ src/attestation/endorser/snp/cert.rs
@@ -78,27 +78,6 @@ impl<'a: 'b, 'b> From<&'a Certificate> for &'b X509 {
     }
 }
 
-/// Verify if the public key of one Certificate signs another Certificate.
-impl Verifiable for (&Certificate, &Certificate) {
-    type Output = ();
-
-    fn verify(self) -> Result<Self::Output> {
-        let signer: X509 = self.0.into();
-        let signee: X509 = self.1.into();
-
-        let key: PKey<Public> = signer.public_key()?;
-        let signed = signee.verify(&key)?;
-
-        match signed {
-            true => Ok(()),
-            false => Err(Error::new(
-                ErrorKind::Other,
-                "Signer certificate does not sign signee certificate",
-            )),
-        }
-    }
-}
-
 impl Certificate {
     /// Create a Certificate from a PEM-encoded X509 structure.
     pub fn from_pem(pem: &[u8]) -> Result<Self> {