+++ src/attestation/verifier/snp/mod.rs
@@ -1,5 +1,16 @@
 // SPDX-License-Identifier: Apache-2.0
 
-//! SNP attestation report verification.
+//! SNP attestation verification.
 
+mod chain;
+mod ecdsa;
 mod report;
+mod signature;
+
+#[cfg(feature = "openssl")]
+mod cert;
+#[cfg(feature = "openssl")]
+mod openssl;
+
+#[cfg(feature = "crypto_nossl")]
+mod cert_nossl;